Bundle similar output filters
The SetNameId and AddEduPersonTargetedId output filters perform a very
similar task. Both filters resolve the NameId, both set an attribute on
the assertion. So bundling them makes sense.

See the final paragraph of: https://github.com/OpenConext/OpenConext-engineblock/pull/1175#pullrequestreview-885974562
